requirements.txt
feedparser>=6.0.12
google-generativeai>=0.8.5
requests>=2.32.5
schedule>=1.2.2
APScheduler>=3.11.1
legendary-gl>=0.20.34
pystray>=0.19.5
pillow>=12.0.0
PyQt6>=6.10.0